+++ backend/woof-connect-api-v1/src/main/kotlin/ru/nnsh/woof_connect/mappers/MappersV1FromTransport.kt
@@ -12,13 +12,13 @@ import ru.nnsh.woof_connect.api.v1.models.DogProfileRequestDebugStubs
 import ru.nnsh.woof_connect.api.v1.models.DogProfileUpdateRequest
 import ru.nnsh.woof_connect.api.v1.models.UserDogIdsRequest
 import ru.nnsh.woof_connect.api.v1.models.UserId
-import ru.nnsh.woof_connect.common.models.WfcContext
-import ru.nnsh.woof_connect.common.models.WfcWorkMode
-import ru.nnsh.woof_connect.common.models.dog_profile.WfcDogId
-import ru.nnsh.woof_connect.common.models.dog_profile.WfcDogProfileBase
-import ru.nnsh.woof_connect.common.models.dog_profile.WfcDogProfileCommand
-import ru.nnsh.woof_connect.common.models.dog_profile.WfcDogProfileStub
-import ru.nnsh.woof_connect.common.models.dog_profile.WfcOwnerId
+import ru.nnsh.woof_connect.common.WfcContext
+import ru.nnsh.woof_connect.common.WfcWorkMode
+import ru.nnsh.woof_connect.common.dog_profile.WfcDogId
+import ru.nnsh.woof_connect.common.dog_profile.WfcDogProfileBase
+import ru.nnsh.woof_connect.common.dog_profile.WfcDogProfileCommand
+import ru.nnsh.woof_connect.common.dog_profile.WfcDogProfileStub
+import ru.nnsh.woof_connect.common.dog_profile.WfcOwnerId
 
 fun WfcContext.fromTransport(request: BaseRequest) = apply {
     when (request) {
