+++ backend/woof-connect-common/src/main/kotlin/ru/nnsh/woof_connect/common/WfcWorkMode.kt
@@ -1,4 +1,4 @@
-package ru.nnsh.woof_connect.common.models
+package ru.nnsh.woof_connect.common
 
 enum class WfcWorkMode {
     PROD, TEST, STUB
