+++ backend/woof-connect-common/src/main/kotlin/ru/nnsh/woof_connect/common/dog_profile/WfcDogProfileStub.kt
@@ -1,4 +1,4 @@
-package ru.nnsh.woof_connect.common.models.dog_profile
+package ru.nnsh.woof_connect.common.dog_profile
 
 enum class WfcDogProfileStub {
     NONE,
